EO-800 Moved DisplayDate into its own component
